pyproject.toml
[project]
name = "primal_dual_ilqr"
version = "0.1"
requires-python = ">= 3.12"
dependencies = [
  "absl-py",
  "jax[cpu]; platform_system != \"Linux\"",
  "jax[cuda12]; platform_system == \"Linux\"",
  "trajax @ git+https://github.com/google/trajax",
]
authors = [
  {name = "Joao Sousa-Pinto", email = "joaospinto@gmail.com"},
]
description = "Implements https://arxiv.org/pdf/2403.00748"
license = "Apache-2.0"

[tool.black]
line-length = 80

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
